% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-good.R
\docType{package}
\name{exampleGood-package}
\alias{exampleGood}
\alias{exampleGood-package}
\title{Example Good}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
