% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{reexport_help}
\alias{reexport_help}
\title{Reexport utils::help}
\usage{
reexport_help(
  topic,
  package = NULL,
  lib.loc = NULL,
  verbose = getOption("verbose"),
  try.all.packages = getOption("help.try.all.packages"),
  help_type = getOption("help_type")
)
}
\description{
Reexport utils::help
}
